refactor: extract datatable pagination logic and types for a better structure
